Fix node test's name First parameter test
